refactor: refactor fizz buzz
Move checking a single round to a helper method
